Add "sensible" madvise calls.
Essentially the strategy is: A newly created and a newly opened file
is marked as "SEQUENTIAL ACCESS", because we will either write to it
or scan it sequentially. As soon as it is sealed, we switch the access
mode to "RANDOM ACCESS", because this should be the normal pattern and
aggressive read-aheads tend to be bad. The collector and the compactor
switch a sealed file back to "SEQUENTIAL ACCESS" just before they scan
it and back to "RANDOM ACCESS", when they are done.

Furthermore, all data files in a collection are advised with "WILLNEED"
just before the collection is scanned during loading.